% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_image.R
\name{arrow_image}
\alias{arrow_image}
\title{Chooses arrow image for first page of bulletin}
\usage{
arrow_image(x)
}
\arguments{
\item{x}{Direction of arrow (up, down, nochange)}
}
\description{
Chooses arrow image for first page of bulletin
}
\examples{
arrow_image("up")

}
