% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.R
\name{crest}
\alias{crest}
\title{Transforms RDOS into tidy format, including offence groups}
\usage{
crest(data)
}
\arguments{
\item{data}{crest_with_offence_groups dataset}
}
\description{
Transforms RDOS into tidy format, including offence groups
}
