% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest_with_offence_groups.R
\name{crest_with_offence_groups}
\alias{crest_with_offence_groups}
\title{Transforms RDOS into tidy format, including offence groups}
\usage{
crest_with_offence_groups(data)
}
\arguments{
\item{data}{CREST RDOS dataset}
}
\description{
Transforms RDOS into tidy format, including offence groups
}
