% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown.R
\name{crown}
\alias{crown}
\title{Returns value of chosen variable at chosen time}
\usage{
crown(df, x, y = 0)
}
\arguments{
\item{df}{CREST dataset}

\item{x}{A variable (disposals, receipts or outstanding).}

\item{y}{A number to indicate time (0 = latest, 1 = one quarter ago, etc.)}
}
\description{
Returns value of chosen variable at chosen time
}
