% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_caseload_data.R
\name{crown_caseload_data}
\alias{crown_caseload_data}
\title{Data integrity checks for table C1 (Crown Court caseload)}
\usage{
crown_caseload_data(x)
}
\arguments{
\item{x}{Table C1.}
}
\value{
If the data is correct, nothing is returned.
}
\description{
Data integrity checks for table C1 (Crown Court caseload)
}
\examples{
crown_caseload_data(CREST_2016)
}
