% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.R
\name{fig2}
\alias{fig2}
\title{Returns Figure 2 (Total Receipts, Disposals Outstanding) in the Criminal Court Quarterly Bulletin}
\usage{
fig2(data)
}
\arguments{
\item{data}{CREST RDOS dataset}
}
\description{
Returns Figure 2 (Total Receipts, Disposals Outstanding) in the Criminal Court Quarterly Bulletin
}
