% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig3.R
\name{fig3}
\alias{fig3}
\title{Returns Figure 3 (Receipt, Disposals, Outstanding by offence group) in the Criminal Court Quarterly Bulletin}
\usage{
fig3(data)
}
\arguments{
\item{data}{CREST RDOS dataset}
}
\description{
Returns Figure 3 (Receipt, Disposals, Outstanding by offence group) in the Criminal Court Quarterly Bulletin
}
