% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig3_data.R
\name{fig3_data}
\alias{fig3_data}
\title{Filters out data after publication date in bulletin}
\usage{
fig3_data(data, publication_year, publication_quarter)
}
\arguments{
\item{data}{CREST RDOS dataset}

\item{publication_year}{Year of publication (integer)}

\item{publication_quarter}{Quarter of publication (integer)}
}
\description{
Filters out data after publication date in bulletin
}
