% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig4.R
\name{fig4}
\alias{fig4}
\title{Returns Figure 4 (Timeliness) in the Criminal Court Quarterly Bulletin}
\usage{
fig4(data)
}
\arguments{
\item{data}{Simplified CREST flatfile dataset}
}
\description{
Returns Figure 4 (Timeliness) in the Criminal Court Quarterly Bulletin
}
