% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{filter_dates}
\alias{filter_dates}
\title{Filters out data after publication date in bulletin}
\usage{
filter_dates(data, publication_year, publication_quarter)
}
\arguments{
\item{data}{CREST RDOS dataset}

\item{publication_year}{Year of publication (integer)}

\item{publication_quarter}{Quarter of publication (integer)}
}
\description{
Filters out data after publication date in bulletin
}
