% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCrest.R
\name{tidyCrest}
\alias{tidyCrest}
\title{Tidy CREST data ready for stats creation.}
\usage{
tidyCrest(x)
}
\arguments{
\item{x}{untidy CREST dataframe}
}
\value{
Tidy CREST dataframe
}
\description{
Tidy CREST data ready for stats creation.
}
\examples{
tidyCrest(df)
}
