% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeliness_data.R
\name{timeliness_data}
\alias{timeliness_data}
\title{Prepares Timeliness data for Section 4 of bulletin}
\usage{
timeliness_data(data, publication_year, publication_quarter)
}
\arguments{
\item{data}{Simplified Crest Flatfile}

\item{publication_year}{Year of publication, integer}

\item{publication_quarter}{Quarter of Publication, integer}
}
\description{
Prepares Timeliness data for Section 4 of bulletin
}
